Split selection tag mutation modules
